backend/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
sentence-transformers
python-multipart
pypdf
python-dotenv
httpx
pydantic
